Support input style environment
